apps/examples : Move 4-channel recording tool to dump4ch. & Restore wakerec.
