+++ resource/kernel/audio/dsp_fw_4ch
@@ -1,0 +1,1 @@
+dump
