+++ resource/kernel/audio/dumptool.ini
@@ -1,5 +1,5 @@
 [config]
 ssid=your_ssid
 psk=your_psk
-server=192.168.68.109
+server=server_ip
 port=12345
